requirements.txt
Django==3.1.6
django-crum==0.7.5
django-widget-tweaks==1.4.8
psycopg2==2.8.5
xhtml2pdf==0.2.4
